--- lib/models/example_user.dart
@@ -1,1 +1,0 @@
-// Define a class representing a user here
