--- lib/providers/example_provider.dart
@@ -1,1 +1,0 @@
-// Put the provider here
