--- lib/services/realtimeDB.dart
@@ -1,1 +1,0 @@
-// Put firebase realtime database service here
