+++ lib/ui/auth/authentication.dart
@@ -2,16 +2,10 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:yiddishconnect/screens/dev_signin_signup/dev_signin.dart';
-import 'package:yiddishconnect/screens/dev_signin_signup/dev_signup.dart';
-import 'package:yiddishconnect/screens/email/emailSignIn.dart';
-import 'package:yiddishconnect/screens/email/emailSignUp.dart';
-import 'package:yiddishconnect/screens/phone/phoneAuth.dart';
-import 'package:yiddishconnect/services/auth.dart';
+import 'package:yiddishconnect/services/firebaseAuthentication.dart';
 import 'package:yiddishconnect/utils/helpers.dart';
-import '../widgets/yd_animated_curve.dart';
-import '../widgets/yd_label.dart';
-import 'dev_signin_signup/dev_home.dart';
+import '../../widgets/yd_animated_curve.dart';
+import '../../widgets/yd_label.dart';
 
 class AuthScreen extends StatelessWidget {
   final AuthService _auth = AuthService();
