+++ lib/ui/auth/email/emailSignIn.dart
@@ -3,7 +3,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:yiddishconnect/utils/helpers.dart';
-import '../../services/auth.dart';
+import '../../../services/firebaseAuthentication.dart';
 import '../dev_signin_signup/dev_home.dart';
 
 class EmailSignInScreen extends StatefulWidget {
