+++ lib/ui/home/home.dart
@@ -5,11 +5,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
-import 'package:yiddishconnect/screens/dev_signin_signup/dev_home.dart';
-import 'package:yiddishconnect/screens/event/eventPage.dart';
-import 'package:yiddishconnect/screens/match/bottomFilter.dart';
-import 'package:yiddishconnect/screens/match/matchPage.dart';
-import 'package:yiddishconnect/screens/match/matchPageProvider.dart';
+import 'package:yiddishconnect/ui/auth/dev_signin_signup/dev_home.dart';
+import 'package:yiddishconnect/ui/home/event/eventPage.dart';
+import 'package:yiddishconnect/ui/home/match/bottomFilter.dart';
+import 'package:yiddishconnect/ui/home/match/matchPage.dart';
+import 'package:yiddishconnect/ui/home/match/matchPageProvider.dart';
 import 'package:yiddishconnect/utils/helpers.dart';
 
 class HomeScreen extends StatefulWidget {
