--- lib/widgets/example_custom_card.dart
@@ -1,1 +1,0 @@
-// This is an example of a custom reusable card widget
